% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kimura_function.R
\name{kimura_length}
\alias{kimura_length}
\title{Kimura Root Length}
\usage{
kimura_length(image)
}
\arguments{
\item{image}{binarized image with root pixels = 1 and background = 0}
}
\value{
root length in px
}
\description{
Estimates the root length. Adapted from Kimura, K., Kikuchi, S., & Yamasaki, S. I. (1999). Accurate root length measurement by image analysis. Plant and Soil, 216(1), 117-127.
}
\examples{
kimura_length(x) = 20
}
