% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png2tiff_function.R
\name{png2tiff}
\alias{png2tiff}
\title{PNG2TIFF file translation}
\usage{
png2tiff(path, path.out)
}
\arguments{
\item{path}{path wHERE png files are expected}

\item{path.out}{path where tiff files going to be stored}
}
\value{
.tiff file
}
\description{
Reads .png images and writes them as .tiff files elsewhere
}
\examples{
png2tiff(image.png) = image.tiff
}
